+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This changelog references the relevant changes done between versions.
 To get the diff for a specific change, go to https://github.com/LIN3S/SharedKernel/commit/XXX where XXX is the change hash 
 To get the diff between two versions, go to https://github.com/LIN3S/SharedKernel/compare/v0.2.0...v0.3.0
 
+* 0.4.0
+    * Added locale value object.
+    * Added zip code and phone value objects.
 * 0.3.0
     * Fixed Value objects' mapping.
     * Added Lin3sSharedKernelBundle.